pyproject.toml
[project]
name = "python-po-lint"
version = "0.1.0"
description = "Lint .po translation files for contamination, wrong languages, shifts, and garbled text"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    {name = "PesCheck"},
]
dependencies = [
    "polib>=1.2",
    "fasttext-wheel>=0.9.2",
    "numpy<2",
]

[project.scripts]
po-lint = "po_lint.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/po_lint"]

[dependency-groups]
dev = [
    "pytest>=7.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.4",
]

[tool.ruff]
line-length = 120

[tool.pytest.ini_options]
testpaths = ["tests"]
